Data Processing (const title = blog?.attributeValues?.blogtitle?.value?.header;, const thumbnail = blog?.attributeValues?.blogthumbnail?.value?.downloadLink;):
Inside the component, it accesses properties of the blog prop to extract necessary data:
title: It extracts the blog title from blog?.attributeValues?.blogtitle?.value?.header. This likely retrieves the title of the blog post.
thumbnail: It extracts the thumbnail image URL from blog?.attributeValues?.blogthumbnail?.value?.downloadLink. This is the image that represents the blog post visually.