Return Statement (return (...)):
The Article component returns JSX, describing the UI structure of an individual article.
JSX Elements (<div>, <Image>, <h1>, <Button>):

The JSX returned by Article:
Starts with a <div> element with class name 'flex flex-col gap-3 border border-gray-400 p-2'. This sets up a flexible container with padding and a border, containing the article content.
Inside this <div>, there are two child elements:
A <div> with class name 'cursor-pointer', indicating an area that can be interacted with (likely for clicking or navigating).
Inside this <div>, there's an <Image> component used to display the blog post thumbnail image.
src attribute of <Image> is set to thumbnail, which is the URL of the blog thumbnail image.
alt attribute is set to "blog-image".
width and height attributes are set to 300, defining the dimensions of the image.
className='w-full' sets the image to fill the entire width of its container.
Another <div> with class name 'text-xl font-bold border-t border-b border-black py-2', providing styling for the article title section.
Inside this <div>, there's an <h1> element that displays the title, which is the extracted title of the blog post.
A <Button> component with className='float-right my-2' and variant={'secondary'} prop.
The button content is "Continue reading ", followed by the ArrowRight icon (<ArrowRight/>) imported earlier.
This button likely serves as a link or action button to continue reading the article.