Return Statement (return (...)):
The ContactButton component returns JSX, describing the UI structure.
JSX Elements (<div>, <Button>, <FormDialog>):

The JSX returned by ContactButton:
Starts with a fragment (<>...</>), allowing multiple child elements without an additional DOM node.
Inside the fragment, there's a <div> with class name 'md:mx-0 mx-2'. This sets up a responsive margin for the button component.
The <div> contains a <Button> component imported earlier from '@/components/ui/button'.
Button has an onClick event handler set to openHandler, triggering the openHandler function when clicked.
It displays the text "CONTACT US" on the button.
className='rounded-sm' applies a small amount of rounding to the button's corners.
Following the <div>, there's a <FormDialog> component.
It's conditionally rendered based on the open state.
open is passed as a prop to determine whether the dialog should be visible or hidden.
setOpen is passed as a prop to allow the dialog to update the open state when closed or submitted.