Imports (import React from 'react', import Articles from '../(components)/Articles'):
React is imported from 'react'. This is necessary for writing React components and JSX.
Articles is imported from '../(components)/Articles'. This imports the Articles component from a relative path, likely from a directory named (components) (the actual directory name might be different, but it's represented here as (components) due to the provided code).